+++ src/main/java/faang/school/projectservice/controller/subproject/ProjectController.java
@@ -9,15 +9,12 @@ import faang.school.projectservice.model.Project;
 import faang.school.projectservice.service.ProjectService;
 import faang.school.projectservice.validator.ProjectValidator;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -54,13 +51,9 @@ public class ProjectController {
         return ResponseEntity.ok(updateProject);
     }
 
-    @GetMapping("/filteredName/{currentUserId}")
+    @PostMapping("/filteredName/{currentUserId}")
     public ResponseEntity<List<ProjectDto>> getProjectsByFilterName(
-            @Valid @RequestBody ProjectFilterDto filterDto,
-            @Min(1)
-            @NotNull
-            @PathVariable
-            Long currentUserId) {
+            @Valid @RequestBody ProjectFilterDto filterDto, @NotNull @PathVariable Long currentUserId) {
         log.info("Getting filtered projects by User #{}.", currentUserId);
 
 
@@ -70,37 +63,25 @@ public class ProjectController {
         return ResponseEntity.ok(projectDtoList);
     }
 
-    @GetMapping("/filteredStatus/{currentUserId}")
+    @PostMapping("/filteredStatus/{currentUserId}")
     public ResponseEntity<List<ProjectDto>> getProjectsByFilterStatus(
-            @Valid @RequestBody ProjectFilterDto filterDto,
-            @Min(1)
-            @NotNull
-            @PathVariable
-            Long currentUserId) {
+            @Valid @RequestBody ProjectFilterDto filterDto, @NotNull @PathVariable Long currentUserId) {
         log.info("Getting filtered projects by User #{}.", currentUserId);
         List<Project> projectsByFilterStatus = projectService.getProjectsByFilterStatus(filterDto, currentUserId);
         List<ProjectDto> projectDtoList = projectMapper.toDtoList(projectsByFilterStatus);
         return ResponseEntity.ok(projectDtoList);
     }
 
-    @GetMapping("/allProjects/{currentUserId}")
+    @PostMapping("/allProjects/{currentUserId}")
     public ResponseEntity<List<Project>> getAllUserAvailableProjects(
-            @Valid @RequestBody ProjectFilterDto filterDto,
-            @Min(1)
-            @NotNull
-            @PathVariable
-            Long currentUserId) {
+            @Valid @RequestBody ProjectFilterDto filterDto, @PathVariable @NotNull Long currentUserId) {
         log.info("Getting available projects by User #{}.", currentUserId);
         return ResponseEntity.ok(projectService.getAllUserAvailableProjects(currentUserId));
     }
 
-    @GetMapping("/filtered/{id}")
+    @PostMapping("/filtered/{id}")
     public ResponseEntity<Project> findProjectById(
-            @Valid @RequestBody ProjectFilterDto filterDto,
-            @Min(1)
-            @NotNull
-            @PathVariable
-            Long id) {
+            @Valid @RequestBody ProjectFilterDto filterDto, @PathVariable @NotNull Long id) {
         log.info("Getting project with #{}.", id);
         return ResponseEntity.ok(projectService.findProjectById(id));
     }
